requirements.txt
# FastAPI and related dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.9.1  # Added for BaseSettings functionality
starlette==0.36.3

# Playwright for browser automation
playwright==1.42.0

# Google API client libraries
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1

# Utility libraries
python-dotenv==1.0.1
aiofiles==23.2.1
python-multipart==0.0.9
jinja2==3.1.3
httpx==0.28.1  # Required for FastAPI TestClient

# Testing libraries
pytest==7.4.4
pytest-asyncio==0.23.4
